Moved Converter, MinioConnection, RethinkDBConnection
